Upgrade fuelphp to 1.4

+++ fuelphp/fuel/core/classes/image/imagemagick.php
@@ -22,9 +22,9 @@ class Image_Imagemagick extends \Image_Driver
 	protected $size_cache = null;
 	protected $im_path = null;
 
-	public function load($filename, $return_data = false)
+	public function load($filename, $return_data = false, $force_extension = false)
 	{
-		extract(parent::load($filename));
+		extract(parent::load($filename, $return_data, $force_extension));
 
 		$this->clear_sizes();
 		if (empty($this->image_temp))
@@ -87,10 +87,13 @@ class Image_Imagemagick extends \Image_Driver
 
 	protected function _watermark($filename, $position, $padding = 5)
 	{
-		extract(parent::_watermark($filename, $x, $y));
+		$values = parent::_watermark($filename, $position, $padding);
+		if ($values == false)
+		{
+			throw new \InvalidArgumentException("Watermark image not found or invalid filetype.");
+		}
 
-		$image = '"'.$this->image_temp.'"';
-		$filename = '"'.$filename.'"';
+		extract($values);
 		$x >= 0 and $x = '+'.$x;
 		$y >= 0 and $y = '+'.$y;
 
@@ -98,7 +101,7 @@ class Image_Imagemagick extends \Image_Driver
 			'composite',
 			'-compose atop -geometry '.$x.$y.' '.
 			'-dissolve '.$this->config['watermark_alpha'].'% '.
-			$filename.' '.$image.' '.$image
+			'"'.$filename.'" "'.$this->image_temp.'" '.$image
 		);
 	}
 
@@ -192,15 +195,15 @@ class Image_Imagemagick extends \Image_Driver
 
 		$this->run_queue();
 		$this->add_background();
-		
+
 		$filetype = $this->image_extension;
 		$old = '"'.$this->image_temp.'"';
 		$new = '"'.$filename.'"';
-		
+
 		if(($filetype == 'jpeg' or $filetype == 'jpg') and $this->config['quality'] != 100)
 		{
 			$quality = '"'.$this->config['quality'].'%"';
-			$this->exec('convert', $old.' -quality '.$quality.' '.strtolower($filetype).' '.$new);
+			$this->exec('convert', $old.' -quality '.$quality.' '.$new);
 		}
 		else
 		{
@@ -221,9 +224,9 @@ class Image_Imagemagick extends \Image_Driver
 
 		$this->run_queue();
 		$this->add_background();
-		
+
 		$image = '"'.$this->image_temp.'"';
-		
+
 		if(($filetype == 'jpeg' or $filetype == 'jpg') and $this->config['quality'] != 100)
 		{
 			$quality = '"'.$this->config['quality'].'%"';
@@ -319,13 +322,16 @@ class Image_Imagemagick extends \Image_Driver
 	 */
 	protected function create_color($hex, $alpha)
 	{
-		export($this->create_hex_color($hex));
+		extract($this->create_hex_color($hex));
 		return "\"rgba(".$red.", ".$green.", ".$blue.", ".round($alpha / 100, 2).")\"";
 	}
 
 	public function __destruct()
 	{
-		unlink($this->image_temp);
+		if (file_exists($this->image_temp))
+		{
+			unlink($this->image_temp);
+		}
 	}
 }
